# config/openapi.yaml
openapi: 3.0.3
info:
  title: ZBI Controller - OpenAPI 3.0
  description: |-
    This is
  termsOfService: https://www.zbitech.net/terms
  contact:
    email: info@zbitech.net
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.1
...
servers:
  - url: https://apps.zbitech.local/api/v1
...
paths:
  /projects:
    post:
      tags:
        - project
      summary: Create a new zbi project
      description: Create a new zbi project
      operationId: createProject
      requestBody:
        description: Create a new zbi project
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                  description: name of project
                network:
                  type: string
                  description: network for the project
                  enum:
                    - testnet
                    - mainnet
                    - regnet
                teamId:
                  type: string
                  description: team identifier for the project
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized request
        '403':
          description: Forbidden
        '429':
          description: Too many requests
        '500':
          description: Internal Server Error
        '503':
          description: Service Unavailable
      security:
        - bearerAuth: []
  /projects/{projectId}:
    get:
      tags:
        - project
      summary: Retrieve a zbi project
      description: Retrieve a zbi project
      operationId: retrieveProject
      parameters:
        - name: projectId
          in: path
          description: id of project
          required: true
          explode: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized request
        '403':
          description: Forbidden
        '429':
          description: Too many requests
        '500':
          description: Internal Server Error
        '503':
          description: Service Unavailable
      security:
        - bearerAuth: []
    delete:
      tags:
        - project
      summary: Delete a zbi project
      description: Delete a zbi project
      operationId: deleteProject
      parameters:
        - name: projectId
          in: path
          description: id of project
          required: true
          explode: true
          schema:
            type: string
      responses:
        '204':
          description: Project description
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized request
        '403':
          description: Forbidden
        '429':
          description: Too many requests
        '500':
          description: Internal Server Error
        '503':
          description: Service Unavailable
      security:
        - bearerAuth: []
  /projects/{projectId}/repair:
    post:
      tags:
        - project
      summary: Repair a zbi project
      description: Repair a zbi project
      operationId: repairProject
      parameters:
        - name: projectId
          in: path
          description: id of project
          required: true
          explode: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized request
        '403':
          description: Forbidden
        '429':
          description: Too many requests
        '500':
          description: Internal Server Error
        '503':
          description: Service Unavailable
      security:
        - bearerAuth: []
...
components:
  schemas:
    Project:
      type: object
      properties:
        id:
          type: string
          description: Project identifier
          example: 83ffec9b-ec1e-494e-8973-ff43171c2c23
        name:
          type: string
          description: Project name
          example: test
        network:
          type: string
          description: network type
          example: testnet
          enum:
            - testnet
            - mainnet
            - regnet
        owner:
          type: string
          description: uid of the project owner
          example: 83ffec9b-ec1e-494e-8973-ff43171c2c23
        team:
          type: string
          description: uid of the project team
          example: 83ffec9b-ec1e-494e-8973-ff43171c2c23
        status:
          type: string
          description: Status of instance
          enum:
            - new
            - created
            - terminating
            - deleted
...
    FieldError:
      type: object
      properties:
        name:
          type: string
          description: name of field
        code:
          type: string
          description: error code
        error:
          type: string
          description: error message
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        type:
          type: string
        message:
          type: string
        fields:
          type: array
          items:
            $ref: '#/components/schemas/ArrayOfFieldErrors'
...
    ArrayOfFieldErrors:
      type: array
      items:
        $ref: '#/components/schemas/FieldError'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT